feat: Add support for llama.cpp embedding provider
- Introduced `EmbeddingProvider` enum to handle multiple embedding providers (Ollama and llama.cpp).
- Updated configuration to include provider selection and corresponding URLs.
- Created `EmbeddingClient` abstraction to dispatch embedding requests to the appropriate provider.
- Implemented `LlamaCppClient` for handling requests to the llama.cpp server.
- Modified worker logic to accommodate the new embedding provider, including job processing and state management.
- Updated README to reflect changes in configuration and operational details.
- Adjusted database interactions to ensure compatibility with the new embedding model.

+++ services/vec-worker/src/db.rs
@@ -367,10 +367,11 @@ pub async fn mark_worker_state(
 /// Release any leases that have expired in the database.
 ///
 /// Calls `SELECT vec_release_expired_leases()` which returns the number of
-/// jobs that were reset to `pending`.
+/// jobs that were reset to `pending`.  The PostgreSQL function returns
+/// `integer`, so we decode into `i32`.
 #[instrument(skip(pool))]
-pub async fn release_expired_leases(pool: &PgPool) -> Result<i64, sqlx::Error> {
-    let row: (i64,) = sqlx::query_as("SELECT vec_release_expired_leases()")
+pub async fn release_expired_leases(pool: &PgPool) -> Result<i32, sqlx::Error> {
+    let row: (i32,) = sqlx::query_as("SELECT vec_release_expired_leases()")
         .fetch_one(pool)
         .await?;
